feat: main page with feature toggling
This makes our main page where all features would be loaded. Features
are stored in an array of JSX Elements displayed upon selection.
Function getClassName is unused right now. It would used in the future
when features have to be placed. List of features would have buttons
which would require className as the return value of this function.